requirements.txt
absl-py==2.2.2
alabaster==0.7.8
apturl==0.5.2
asciitree==0.3.3
astunparse==1.6.3
atomicwrites==1.1.5
attrs==19.3.0
Automat==0.8.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.8.2
blosc==1.7.0
Bottleneck==1.2.1
brial==1.0.2
Brlapi==0.7.0
cachetools==5.5.2
cdo==1.5.3
certifi==2019.11.28
cftime==1.1.0
chardet==3.0.4
click==8.1.8
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cvxopt==1.2.3
cycler==0.10.0
cypari2==2.1.1
Cython==0.29.14
dask==2.8.1+dfsg
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.16.1
flask==3.0.3
flatbuffers==25.2.10
fpylll==0.5.1.dev0
fsspec==0.6.1
future==0.18.2
gast==0.4.0
gmpy2==2.1.0b3
google-auth==2.40.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5netcdf==0.7.1
h5py==3.11.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==8.5.0
incremental==16.10.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.2.0
jdcal==1.0
jedi==0.15.2
jinja2==3.1.6
joblib==1.4.2
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
keras==2.13.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
locket==0.2.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.1.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.1.0
msgpack==0.6.2
nbconvert==5.6.1
nbformat==5.0.4
netCDF4==1.5.3
netifaces==0.10.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numcodecs==0.0.0
numexpr==2.7.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
opencv-python==4.11.0.86
openpyxl==3.0.3
opt-einsum==3.4.0
packaging==20.3
pandas==2.0.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
partd==1.0.0
path.py==12.0.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkgconfig==1.5.1
pluggy==0.13.0
pplpy==0.8.4
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==4.25.7
psutil==5.5.1
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
PyStemmer==1.3.0
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
python3-openid==3.1.0
pytz==2025.2
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
roman==2.0.0
rpy2==3.2.6
rsa==4.9.1
sage==9.0
sagenb-export==3.2
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
setuptools-scm==3.4.3
simplegeneric==0.8.1
simplejson==3.16.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==1.9.5
Sphinx==1.8.5
sympy==1.5.1
systemd-python==234
tables==3.6.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.8.2
testpath==0.4.4
toml==0.10.0
toolz==0.9.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==5.1.1
traitlets==4.3.3
triton==3.0.0
Twisted==18.9.0
typing-extensions==4.13.2
tzdata==2025.2
tzlocal==2.1b1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vcversioner==2.16.0.0
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
werkzeug==3.0.6
widgetsnbextension==2.0.0
wrapt==1.17.2
xarray==0.15.0
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
zarr==0.0.0
zipp==3.20.2
zope.interface==4.7.1
